mod[api]: extract shared mic-input helper for examples (QVAC-13760)
Deduplicate the ffmpeg microphone setup that was copy-pasted across the
voice-assistant, whispercpp-microphone-record, and
parakeet-microphone-record examples. New shared helper lives at
examples/audio/mic-input.ts and exposes startMicrophone({ sampleRate,
format }), getAudioInputArgs(), and listWindowsAudioDevices().

All three examples now import from the shared helper, removing ~230
lines of duplication.
