+++ packages/sdk/examples/transcription/parakeet-microphone-record.ts
@@ -17,65 +17,13 @@ import {
   PARAKEET_TDT_VOCAB,
   PARAKEET_TDT_PREPROCESSOR_FP32,
 } from "@qvac/sdk";
-import { spawn, spawnSync } from "child_process";
-import { platform } from "os";
+import { spawnSync } from "child_process";
+import { startMicrophone } from "../audio/mic-input";
 
 const SAMPLE_RATE = 16000;
 const BYTES_PER_SAMPLE = 2; // s16le
 const CHUNK_DURATION_S = 3;
 const CHUNK_SIZE = SAMPLE_RATE * BYTES_PER_SAMPLE * CHUNK_DURATION_S;
-
-function listWindowsAudioDevices(): string[] {
-  const result = spawnSync(
-    "ffmpeg",
-    ["-hide_banner", "-list_devices", "true", "-f", "dshow", "-i", "dummy"],
-    { encoding: "utf8" },
-  );
-  const stderr = result.stderr ?? "";
-  const devices: string[] = [];
-  let inAudioSection = false;
-  for (const line of stderr.split(/\r?\n/)) {
-    if (line.includes("DirectShow audio devices")) {
-      inAudioSection = true;
-      continue;
-    }
-    if (line.includes("DirectShow video devices")) {
-      inAudioSection = false;
-      continue;
-    }
-    if (!inAudioSection) continue;
-    const match = line.match(/"([^"]+)"/);
-    if (!match) continue;
-    const name = match[1];
-    if (!name || name.startsWith("@device")) continue;
-    devices.push(name);
-  }
-  return devices;
-}
-
-function getAudioInputArgs(): string[] {
-  const env = process.env as Record<string, string | undefined>;
-  const override = env["MIC_DEVICE"];
-  switch (platform()) {
-    case "darwin":
-      return ["-f", "avfoundation", "-i", override ?? ":0"];
-    case "linux":
-      return ["-f", "pulse", "-i", override ?? "default"];
-    case "win32": {
-      const deviceName = override ?? listWindowsAudioDevices()[0];
-      if (!deviceName) {
-        throw new Error(
-          "No Windows audio input device found. List devices with:\n" +
-            "  ffmpeg -hide_banner -list_devices true -f dshow -i dummy\n" +
-            'Then set MIC_DEVICE="Your Device Name" and retry.',
-        );
-      }
-      return ["-f", "dshow", "-i", `audio=${deviceName}`];
-    }
-    default:
-      throw new Error(`Unsupported platform: ${platform()}`);
-  }
-}
 
 // ── Main ──
 
@@ -101,23 +49,10 @@ const modelId = await loadModel({
 });
 console.log("Model loaded.\n");
 
-const ffmpeg = spawn(
-  "ffmpeg",
-  [
-    ...getAudioInputArgs(),
-    "-ar",
-    String(SAMPLE_RATE),
-    "-ac",
-    "1",
-    "-sample_fmt",
-    "s16",
-    "-f",
-    "s16le",
-    "pipe:1",
-  ],
-  { stdio: ["ignore", "pipe", "ignore"] },
-);
-if (!ffmpeg.stdout) throw new Error("Failed to open microphone");
+const ffmpeg = startMicrophone({
+  sampleRate: SAMPLE_RATE,
+  format: "s16le",
+});
 
 let buffer = Buffer.alloc(0);
 let processing = false;
