+++ packages/sdk/examples/transcription/whispercpp-microphone-record.ts
@@ -15,62 +15,10 @@ import {
   WHISPER_TINY,
   VAD_SILERO_5_1_2,
 } from "@qvac/sdk";
-import { spawn, spawnSync } from "child_process";
-import { platform } from "os";
+import { spawnSync } from "child_process";
+import { startMicrophone } from "../audio/mic-input";
 
 const SAMPLE_RATE = 16000;
-
-function listWindowsAudioDevices(): string[] {
-  const result = spawnSync(
-    "ffmpeg",
-    ["-hide_banner", "-list_devices", "true", "-f", "dshow", "-i", "dummy"],
-    { encoding: "utf8" },
-  );
-  const stderr = result.stderr ?? "";
-  const devices: string[] = [];
-  let inAudioSection = false;
-  for (const line of stderr.split(/\r?\n/)) {
-    if (line.includes("DirectShow audio devices")) {
-      inAudioSection = true;
-      continue;
-    }
-    if (line.includes("DirectShow video devices")) {
-      inAudioSection = false;
-      continue;
-    }
-    if (!inAudioSection) continue;
-    const match = line.match(/"([^"]+)"/);
-    if (!match) continue;
-    const name = match[1];
-    if (!name || name.startsWith("@device")) continue;
-    devices.push(name);
-  }
-  return devices;
-}
-
-function getAudioInputArgs(): string[] {
-  const env = process.env as Record<string, string | undefined>;
-  const override = env["MIC_DEVICE"];
-  switch (platform()) {
-    case "darwin":
-      return ["-f", "avfoundation", "-i", override ?? ":0"];
-    case "linux":
-      return ["-f", "pulse", "-i", override ?? "default"];
-    case "win32": {
-      const deviceName = override ?? listWindowsAudioDevices()[0];
-      if (!deviceName) {
-        throw new Error(
-          "No Windows audio input device found. List devices with:\n" +
-            "  ffmpeg -hide_banner -list_devices true -f dshow -i dummy\n" +
-            'Then set MIC_DEVICE="Your Device Name" and retry.',
-        );
-      }
-      return ["-f", "dshow", "-i", `audio=${deviceName}`];
-    }
-    default:
-      throw new Error(`Unsupported platform: ${platform()}`);
-  }
-}
 
 // ── Main ──
 
@@ -107,23 +55,10 @@ const modelId = await loadModel({
 });
 console.log("Model loaded.\n");
 
-const ffmpeg = spawn(
-  "ffmpeg",
-  [
-    ...getAudioInputArgs(),
-    "-ar",
-    String(SAMPLE_RATE),
-    "-ac",
-    "1",
-    "-sample_fmt",
-    "flt",
-    "-f",
-    "f32le",
-    "pipe:1",
-  ],
-  { stdio: ["ignore", "pipe", "ignore"] },
-);
-if (!ffmpeg.stdout) throw new Error("Failed to open microphone");
+const ffmpeg = startMicrophone({
+  sampleRate: SAMPLE_RATE,
+  format: "f32le",
+});
 
 const session = await transcribeStream({ modelId });
